test: add new scenario tests for user class

+++ src/test/java/it/polito/ezgas/TestScenarios.java
@@ -20,7 +20,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import it.polito.ezgas.converter.UserConverter;
 import it.polito.ezgas.dto.IdPw;
 import it.polito.ezgas.dto.LoginDto;
-
+import it.polito.ezgas.dto.UserDto;
 import it.polito.ezgas.entity.User;
 import it.polito.ezgas.repository.UserRepository;
 import it.polito.ezgas.service.impl.UserServiceimpl;
@@ -50,8 +50,8 @@ public class TestScenarios {
 				User user09= new User("Lisa"    , "Garrison" ,"Garrison@ezgaz.com" ,1 );
 				User user10= new User("Amari"   , "Jenkins"  ,"Jenkins@ezgaz.com"  ,3 );
 				User user11= new User("Jayson"  , "Anderson" ,"Anderson@ezgaz.com" ,1 );
-			  
-				
+			    User user12= new User("Pierre"  , "Cox" ,"Pierre@ezgaz.com" ,1 );
+			    user12.setAdmin(true);
 				 
 				userslist.add(user00);
 				userslist.add(user01);
@@ -65,6 +65,7 @@ public class TestScenarios {
 				userslist.add(user09);
 				userslist.add(user10);
 				userslist.add(user11);
+				userslist.add(user12);
 				
 				userRepository.save(userslist);
 				
@@ -84,25 +85,65 @@ public class TestScenarios {
 			userRepository.delete( userRepository.findByEmail("Shepherd@ezgaz.com")); 
 			userRepository.delete( userRepository.findByEmail("Dorsey@ezgaz.com"  )); 
 			userRepository.delete( userRepository.findByEmail("Paul@ezgaz.com"    )); 
+			if(!userRepository.findByEmail("Buckley@ezgaz.com").isEmpty())
 			userRepository.delete( userRepository.findByEmail("Buckley@ezgaz.com" )); 
 			userRepository.delete( userRepository.findByEmail("Goodwin@ezgaz.com" )); 
 			userRepository.delete( userRepository.findByEmail("Garrison@ezgaz.com")); 
 			userRepository.delete( userRepository.findByEmail("Jenkins@ezgaz.com" )); 
 			userRepository.delete( userRepository.findByEmail("Anderson@ezgaz.com")); 
+			
+			if(!userRepository.findByEmail("Gardner@ezgaz.it").isEmpty())
+				userRepository.delete( userRepository.findByEmail("Gardner@ezgaz.it"));
+			if(!userRepository.findByEmail("Pierre@ezgaz.com").isEmpty())
+				userRepository.delete( userRepository.findByEmail("Pierre@ezgaz.com"));
+		
+			
 		}catch (Exception e) 
 		{ 
 			fail ("EndTest Error");
 		}
     }
-	
 	@Test
 	public void testScenario1() {
 		try {
 			Connection conn = DriverManager.getConnection("jdbc:h2:./data/memo", "sa", "password");
 			UserServiceimpl userImpl = new UserServiceimpl(userRepository,userConverter);
+			UserDto user = new UserDto(1, "Gardner", "Kelvin", "Gardner@ezgaz.it", 1);
+			userImpl.saveUser(user); 
+			
+			
+		   LoginDto userdto= userImpl.login(new IdPw("Gardner@ezgaz.it","Kelvin"));
+		    ArrayList<User> log = userRepository.findByEmail(userdto.getEmail());
+		    
+			if(log.size()==1)
+				{
+				
+				 User foundUser= log.get(0);	 
+				 assertEquals("Gardner",foundUser.getUserName());
+				}
+			    else if(log.size()>0)
+				{
+				 
+				 fail("More than one user found with this email");			
+				}
+			     else if(log.size()==0)
+				{ 
+				 fail("No user found with this email");		
+				}
+			conn.close(); 
+		} catch (Exception e) {
+			fail("Exception throwed");
+			}
+		
+	}
+	@Test
+	public void testScenario2() {
+		try {
+			Connection conn = DriverManager.getConnection("jdbc:h2:./data/memo", "sa", "password");
+			UserServiceimpl userImpl = new UserServiceimpl(userRepository,userConverter);
 		LoginDto userdto=	userImpl.login(new IdPw("Shepherd@ezgaz.com","Shepherd"));
 		    ArrayList<User> log = userRepository.findByEmail(userdto.getEmail());
-		    System.out.println(log.get(0).getUserName());
+		 
 			if(log.size()==1)
 				{
 				
@@ -116,6 +157,45 @@ public class TestScenarios {
 				 Integer NewRep=userImpl.increaseUserReputation(foundUser.getUserId());
 				assertEquals(userRep,NewRep ); 
 				
+				 
+				}
+			    else if(log.size()>0)
+				{
+				 
+				 fail("More than one user found with this email");			
+				}
+			     else if(log.size()==0)
+				{
+				 
+				 fail("No user found with this email");		
+				}
+			conn.close(); 
+		} catch (Exception e) {
+			fail("Exception throwed");
+			}
+		
+	}
+	@Test
+	public void testScenario3() {
+		try {
+			Connection conn = DriverManager.getConnection("jdbc:h2:./data/memo", "sa", "password");
+			UserServiceimpl userImpl = new UserServiceimpl(userRepository,userConverter);
+		    LoginDto userdto=	userImpl.login(new IdPw("Pierre@ezgaz.com","Cox"));
+		    ArrayList<User> log = userRepository.findByEmail(userdto.getEmail());
+		 
+			if(log.size()==1)
+				{
+				
+				 User foundUser= log.get(0);
+				 if(foundUser.getAdmin())
+				 {
+					    ArrayList<User> userclient = userRepository.findByEmail("Buckley@ezgaz.com");
+					    userImpl.deleteUser(userclient.get(0).getUserId());
+				 }
+				 else
+					 
+					 fail("User is not admin to delete the user");
+				 
 				 
 				}
 			    else if(log.size()>0)
